Refactor transcription utils and document modules
