+++ prompt.md
@@ -2,3 +2,5 @@ The initial prompt for this repository was:
 
 "Create a minimal Tauri desktop application that can generate videos from audio, add captions and basic media elements, and upload the result to YouTube. Provide both a GUI built with React/TypeScript and a CLI. Use Rust on the backend to perform the heavy lifting with ffmpeg and whisper for transcription."
 
+Subsequent updates introduced batch processing, subtitle translation utilities and OAuth based YouTube uploads while keeping the architecture lightweight.
+
