+++ prompttracking.md
@@ -1,4 +1,5 @@
 # Prompt History
 
 - Initial prompt: see `prompt.md`.
+- Refactor and document the codebase, adding translation utilities and extensive comments.
