+++ ytapp/src/features/transcription/index.ts
@@ -1,7 +1,7 @@
 // Helpers for generating SRT subtitles with optional translation.
 import { invoke } from '@tauri-apps/api/core';
-import { spawn } from 'child_process';
 import { Language } from '../language';
+import { translateSrt } from '../utils/translate';
 
 export interface TranscribeParams {
     file: string;
@@ -12,21 +12,13 @@ export interface TranscribeParams {
     translate?: string[];
 }
 
-function translateSrt(input: string, target: string): Promise<string> {
-    const output = input.replace(/\.srt$/, `.${target}.srt`);
-    return new Promise((resolve, reject) => {
-        const child = spawn('argos-translate', [
-            '--input-file', input,
-            '--output-file', output,
-            '--from-lang', 'en',
-            '--to-lang', target,
-        ]);
-        child.on('exit', code => {
-            if (code === 0) resolve(output); else reject(new Error('translation failed'));
-        });
-    });
-}
-
+/**
+ * Transcribe an audio file using the backend and optionally translate the
+ * resulting subtitles into additional languages.
+ *
+ * @param params Options controlling the transcription and translation.
+ * @returns Array of generated subtitle paths.
+ */
 export async function transcribeAudio(params: TranscribeParams): Promise<string[]> {
     const { file, language = 'auto', translate } = params;
     const result: string = await invoke('transcribe_audio', { file, language });
